requirements.txt
amuse_athena==2023.5.0
amuse_bhtree==2023.5.0
amuse_bse==2023.5.0
amuse_capreole==2023.5.0
amuse_evtwin==2023.5.0
amuse_fastkick==2023.5.0
amuse_fi==2023.5.0
amuse_fractalcluster==2023.5.0
amuse_framework==2023.5.0
amuse_gadget2==2023.5.0
amuse_galactics==2023.5.0
amuse_galaxia==2023.5.0
amuse_halogen==2023.5.0
amuse_hermite==2023.5.0
amuse_hop==2023.5.0
amuse_huayno==2023.5.0
amuse_kepler==2023.5.0
amuse_kepler_orbiters==2023.5.0
amuse_mameclot==2023.5.0
amuse_mercury==2023.5.0
amuse_mesa_r2208==2023.5.0
amuse_mmams==2023.5.0
amuse_ph4==2023.5.0
amuse_seba==2023.5.0
amuse_secularmultiple==2023.5.0
amuse_smalln==2023.5.0
amuse_sphray==2023.5.0
amuse_sse==2023.5.0
amuse_twobody==2023.5.0
matplotlib==3.7.2
numpy==1.25.1
pandas==2.0.3
py_ubjson==0.16.1
scipy==1.11.2
seaborn==0.12.2
tqdm==4.66.0
zstandard==0.21.0
